rep_* functions moved to utils.py
