+++ tests_random/utils.py
@@ -22,6 +22,7 @@
 # IN THE SOFTWARE.
 #-------------------------------------------------------------------------------
 import sys
+import textwrap
 from datatable.internal import frame_integrity_check
 
 
@@ -72,3 +73,72 @@ def assert_equals(frame1, frame2):
         assert same_iterables(frame1.names, frame2.names)
         assert same_iterables(frame1.stypes, frame2.stypes)
         assert same_iterables(frame1.to_list(), frame2.to_list())
+
+
+
+def repr_row(row, j):
+    assert 0 <= j < len(row)
+    if len(row) <= 20 or j <= 12:
+        out = str(row[:j])[:-1]
+        if j > 0:
+            out += ",   " + str(row[j])
+        if len(row) > 20:
+            out += ",   " + str(row[j+1:20])[1:-1] + ", ...]"
+        elif j < len(row) - 1:
+            out += ",   " + str(row[j+1:])[1:]
+        else:
+            out += "]"
+    else:
+        out = str(row[:10])[:-1] + ", ..., " + str(row[j]) + ", ...]"
+    return out
+
+
+
+def repr_slice(s):
+    if s == slice(None):
+        return ":"
+    if s.start is None:
+        if s.stop is None:
+            return "::" +  str(s.step)
+        elif s.step is None:
+            return ":" + str(s.stop)
+        else:
+            return ":" + str(s.stop) + ":" + str(s.step)
+    else:
+        if s.stop is None:
+            return str(s.start) + "::" +  str(s.step)
+        elif s.step is None:
+            return str(s.start) + ":" + str(s.stop)
+        else:
+            return str(s.start) + ":" + str(s.stop) + ":" + str(s.step)
+
+
+def repr_data(data, indent):
+    out = "["
+    for i, arr in enumerate(data):
+        if i > 0:
+            out += " " * (indent + 1)
+        out += "["
+        out += textwrap.fill(", ".join(repr(e) for e in arr),
+                             width = 80, subsequent_indent=" "*(indent+2))
+        out += "]"
+        if i != len(data) - 1:
+            out += ",\n"
+    out += "]"
+    return out
+
+
+def repr_type(t):
+    if t is int:
+        return "int"
+    if t is bool:
+        return "bool"
+    if t is float:
+        return "float"
+    if t is str:
+        return "str"
+    raise ValueError("Unknown type %r" % t)
+
+def repr_types(types):
+    assert isinstance(types, (list, tuple))
+    return "[" + ", ".join(repr_type(t) for t in types) + "]"
